Code cleaning in preparation for #7258 [3/n] (#7262)
